requirements.txt
streamlit>=1.40.0
numpy>=1.26.0
faiss-cpu>=1.8.0
scikit-learn>=1.5.0
pdfplumber>=0.11.0
groq>=0.11.0
openai-whisper>=20240930
